requirements.txt
coverage==7.1.0
psycopg2-binary==2.9.7
